requirements.txt
SQLAlchemy==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
APScheduler==3.10.4
httpx==0.25.2
beautifulsoup4==4.14.3
lxml==5.2.2
playwright==1.41.2